pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "relay-desktop-assistant"
version = "1.0.0"
description = "Universal Desktop Assistant powered by GPT-4V"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "RELAY Team", email = "relay@example.com"}
]
maintainers = [
    {name = "RELAY Team", email = "relay@example.com"}
]
keywords = ["automation", "desktop", "assistant", "vision", "gpt-4", "pyautogui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]
requires-python = ">=3.8.1"
dependencies = [
    "openai>=1.0.0",
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pynput>=1.7.6",
    "keyboard>=0.13.5",
    "mouse>=0.7.1"

]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
relay = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/relay"
Documentation = "https://github.com/yourusername/relay#readme"
Repository = "https://github.com/yourusername/relay"
Issues = "https://github.com/yourusername/relay/issues"

[tool.hatch.build.targets.wheel]
packages = ["relay"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["relay"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "customtkinter.*",
    "pyautogui.*",
    "pynput.*",
    "keyboard.*",
    "mouse.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=relay",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
] 